Direct method of BlockOperator returns a single DataContainer if the output shape is (1,1) (#1802)
Direct method of BlockOperator returns a single DataContainer if the output shape is (1,1) 

---------
