+++ Wrappers/Python/cil/optimisation/operators/BlockOperator.py
@@ -191,11 +191,10 @@ class BlockOperator(Operator):
             x_b = BlockDataContainer(x)
         else:
             x_b = x
-        shape = self.get_output_shape(x_b.shape)
-        res = []
+        shape = self.get_output_shape(x_b.shape)        
 
         if out is None:
-
+            res = []
             for row in range(self.shape[0]):
                 for col in range(self.shape[1]):
                     if col == 0:
@@ -205,7 +204,13 @@ class BlockOperator(Operator):
                         prod += self.get_item(row,
                                               col).direct(x_b.get_item(col))
                 res.append(prod)
-            return BlockDataContainer(*res, shape=shape)
+            if 1 == shape[0] == shape[1]:
+                # the output is a single DataContainer, so we can take it out 
+                return res[0] 
+            else: 
+                return BlockDataContainer(*res, shape=shape) 
+            
+
         else:
             tmp = self.range_geometry().allocate()
             for row in range(self.shape[0]):
@@ -369,6 +374,8 @@ class BlockOperator(Operator):
             tmp = []
             for i in range(self.shape[1]):
                 tmp.append(self.get_item(0, i).domain_geometry())
+            if self.shape[1] == 1:
+                return tmp[0]
             return BlockGeometry(*tmp)
 
             # shape = (self.shape[0], 1)
@@ -381,6 +388,8 @@ class BlockOperator(Operator):
         tmp = []
         for i in range(self.shape[0]):
             tmp.append(self.get_item(i, 0).range_geometry())
+        if self.shape[0] == 1:
+            return tmp[0]
         return BlockGeometry(*tmp)
 
     def sum_abs_row(self):
